refactor instrumented tests
