--- app/src/androidTest/java/com/mobiquity/mobproducts/helper/TestApplication.kt
@@ -1,15 +1,0 @@
-package com.mobiquity.mobproducts.helper
-
-import android.app.Application
-import androidx.fragment.app.Fragment
-import dagger.android.DispatchingAndroidInjector
-import dagger.android.support.HasSupportFragmentInjector
-import javax.inject.Inject
-
-class TestApplication : Application(), HasSupportFragmentInjector {
-
-    @Inject
-    lateinit var fragmentInjector: DispatchingAndroidInjector<Fragment>
-
-    override fun supportFragmentInjector() = fragmentInjector
-}
