+++ app/src/androidTest/java/com/mobiquity/mobproducts/helper/TestInjector.kt
@@ -2,8 +2,8 @@ package com.mobiquity.mobproducts.helper
 
 import androidx.test.platform.app.InstrumentationRegistry
 import com.mobiquity.mobproducts.ProductsApplicaton
-import com.mobiquity.mobproducts.mockedApi.DaggerTestApplicationComponent
-import com.mobiquity.mobproducts.mockedApi.FakeRepositoryModule
+import com.mobiquity.mobproducts.di.DaggerTestApplicationComponent
+import com.mobiquity.mobproducts.di.FakeRepositoryModule
 
 class TestInjector(private val testApplicationModule: FakeRepositoryModule) {
